Move specializations of auto_arg to their respective types
These specializations require access to AutoPacket arguments.  Move them to the types they specialize, where possible, to break a dependency cycle.
