+++ autowiring/autowiring.h
@@ -3,3 +3,6 @@
 
 #include "Autowired.h"
 #include "AutoNetServer.h"
+#include "auto_in.h"
+#include "auto_out.h"
+#include "auto_prev.h"
